# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Telegram-clone chat API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface OtpRequest {
  /** E.164 or local phone number */
  phone: string;
}

export interface OtpRequestResult {
  sessionId: string;
  message: string;
}

export interface OtpVerify {
  sessionId: string;
  code: string;
}

export interface OtpVerifyResult {
  verified: boolean;
  isNewUser: boolean;
  /** Token to pass to setup-profile or login */
  sessionToken: string;
}

export interface ProfileSetup {
  sessionToken: string;
  /** @minLength 3 */
  username: string;
  /** @minLength 6 */
  password: string;
  /** @nullable */
  bio?: string | null;
}

export interface LoginInput {
  /** Phone number or username */
  identifier: string;
  password: string;
}

export interface User {
  id: number;
  phone: string;
  username: string;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
}

export interface AuthResult {
  user: User;
  token: string;
}

export interface ProfileUpdate {
  /** @minLength 3 */
  username?: string;
  /** @nullable */
  bio?: string | null;
  phone?: string;
  /** @minLength 6 */
  password?: string;
  /** @nullable */
  avatarUrl?: string | null;
}

export type ChatType = (typeof ChatType)[keyof typeof ChatType];

export const ChatType = {
  direct: "direct",
  group: "group",
} as const;

export type ChatMemberRole =
  (typeof ChatMemberRole)[keyof typeof ChatMemberRole];

export const ChatMemberRole = {
  admin: "admin",
  member: "member",
} as const;

export interface ChatMember {
  user: User;
  role: ChatMemberRole;
  joinedAt: string;
}

export interface Message {
  id: number;
  chatId: number;
  senderId: number;
  content: string;
  createdAt: string;
  sender?: User;
}

export interface Chat {
  id: number;
  type: ChatType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  members: ChatMember[];
  lastMessage?: Message;
}

export type ChatSummaryType =
  (typeof ChatSummaryType)[keyof typeof ChatSummaryType];

export const ChatSummaryType = {
  direct: "direct",
  group: "group",
} as const;

export interface ChatSummary {
  id: number;
  type: ChatSummaryType;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  createdAt: string;
  lastMessage?: Message;
  unreadCount: number;
  otherUser?: User;
}

export type ChatInputType = (typeof ChatInputType)[keyof typeof ChatInputType];

export const ChatInputType = {
  direct: "direct",
  group: "group",
} as const;

export interface ChatInput {
  type: ChatInputType;
  name?: string;
  description?: string;
  memberIds?: number[];
}

export interface ChatUpdate {
  name?: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
}

export interface MemberInput {
  userId: number;
}

export interface MessageInput {
  /** @minLength 1 */
  content: string;
}

export interface VoiceParticipant {
  user: User;
  micEnabled: boolean;
  cameraEnabled: boolean;
  joinedAt: string;
}

export interface VoiceRoom {
  chatId: number;
  active: boolean;
  participants: VoiceParticipant[];
}

export interface VoiceJoinInput {
  micEnabled: boolean;
  cameraEnabled: boolean;
}

export type SearchUsersParams = {
  /**
   * Phone number or username to search
   */
  q: string;
};

export type GetChatMessagesParams = {
  /**
   * Cursor for pagination (message ID)
   */
  before?: number;
  limit?: number;
};