# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Telegram-clone chat API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/request-otp:
    post:
      operationId: requestOtp
      tags: [auth]
      summary: Request OTP code for phone number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "200":
          description: OTP sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestResult"
        "400":
          description: Invalid phone number
  /auth/verify-otp:
    post:
      operationId: verifyOtp
      tags: [auth]
      summary: Verify OTP code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerify"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpVerifyResult"
        "400":
          description: Invalid or expired OTP
  /auth/setup-profile:
    post:
      operationId: setupProfile
      tags: [auth]
      summary: Complete registration by setting username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileSetup"
      responses:
        "200":
          description: Profile created, user logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "400":
          description: Username taken or invalid data
        "401":
          description: No verified OTP session
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with phone/username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated

  # ── Users ─────────────────────────────────────────────────────────────────
...
  /chats:
    get:
      operationId: listChats
      tags: [chats]
      summary: List all chats for current user
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatSummary"
        "401":
          description: Not authenticated

    post:
      operationId: createChat
      tags: [chats]
      summary: Create a direct chat or group chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatInput"
      responses:
        "201":
          description: Chat created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "401":
          description: Not authenticated
  /chats/{chatId}:
    get:
      operationId: getChat
      tags: [chats]
      summary: Get a specific chat with member details
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
        "404":
          description: Chat not found

    patch:
      operationId: updateChat
      tags: [chats]
      summary: Update group chat details (name, description)
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatUpdate"
      responses:
        "200":
          description: Chat updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{chatId}/members:
    post:
      operationId: addChatMember
      tags: [chats]
      summary: Add member to group chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "200":
          description: Member added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"
  /chats/{chatId}/members/{userId}:
    delete:
      operationId: removeChatMember
      tags: [chats]
      summary: Remove member from group chat
      parameters:
        - name: chatId
          in: path
          required: true
          schema:
            type: integer
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Member removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /chats/direct/{userId}:
    post:
      operationId: openDirectChat
      tags: [chats]
      summary: Open or get an existing direct chat with a user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Direct chat
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Chat"

  # ── Messages ──────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    # ── Auth schemas ──────────────────────────────────────────────────────
    OtpRequest:
      type: object
      required: [phone]
      properties:
        phone:
          type: string
          description: E.164 or local phone number
    OtpRequestResult:
      type: object
      required: [sessionId, message]
      properties:
        sessionId:
          type: string
        message:
          type: string
    OtpVerify:
      type: object
      required: [sessionId, code]
      properties:
        sessionId:
          type: string
        code:
          type: string
    OtpVerifyResult:
      type: object
      required: [verified, isNewUser, sessionToken]
      properties:
        verified:
          type: boolean
        isNewUser:
          type: boolean
        sessionToken:
          type: string
          description: Token to pass to setup-profile or login
    ProfileSetup:
      type: object
      required: [sessionToken, username, password]
      properties:
        sessionToken:
          type: string
        username:
          type: string
          minLength: 3
        password:
          type: string
          minLength: 6
        bio:
          type: ["string", "null"]
    LoginInput:
      type: object
      required: [identifier, password]
      properties:
        identifier:
          type: string
          description: Phone number or username
        password:
          type: string
    AuthResult:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string

    # ── User schemas ──────────────────────────────────────────────────────
    User:
      type: object
      required: [id, phone, username, createdAt]
      properties:
        id:
          type: integer
        phone:
          type: string
        username:
          type: string
        bio:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Chat:
      type: object
      required: [id, type, createdAt, members]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group]
        name:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        members:
          type: array
          items:
            $ref: "#/components/schemas/ChatMember"
        lastMessage:
          $ref: "#/components/schemas/Message"
    ChatSummary:
      type: object
      required: [id, type, createdAt, unreadCount]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [direct, group]
        name:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastMessage:
          $ref: "#/components/schemas/Message"
        unreadCount:
          type: integer
        otherUser:
          $ref: "#/components/schemas/User"
    ChatInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [direct, group]
        name:
          type: string
        description:
          type: string
        memberIds:
          type: array
          items:
            type: integer
    ChatUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
    ChatMember:
      type: object
      required: [user, role, joinedAt]
      properties:
        user:
          $ref: "#/components/schemas/User"
        role:
          type: string
          enum: [admin, member]
        joinedAt:
          type: string
          format: date-time
    MemberInput:
      type: object
      required: [userId]
      properties:
        userId:
          type: integer

    # ── Message schemas ───────────────────────────────────────────────────
    Message:
      type: object
      required: [id, chatId, senderId, content, createdAt]
      properties:
        id:
          type: integer
        chatId:
          type: integer
        senderId:
          type: integer
        content:
          type: string
        createdAt:
          type: string
          format: date-time
        sender:
          $ref: "#/components/schemas/User"
...